==> 01_data_capture.ipynb <==
import json
import requests
import re
from os.path import join
from datetime import datetime
import progressbar

# Parsing of post
def parse_post(post):
    output_post = {}
    output_post['id'] = post['id']
    output_post['url'] = post['url']
    output_post['dateRFC'] = post['dateRFC']
    output_post['date'] = post['date']
    output_post['title'] = post['title']
    output_post['intro'] = post['intro']
    text = ''
    
    # Собираем текст в одно целое
    for block in post['blocks']:
        if block['type'] == 'text' or block['type'] == 'quote':
            text += block['data']['text'] + ' '
        if block['type'] == 'list':
            text += ' '.join(block['data']['items']) + ' '
        if block['type'] == 'rawhtml':
            text += cleanhtml(block['data']['raw']) + ' '
            
    # Удаляем штуки которые мешают читать текст
    text = text.replace('[','', -1)
    text = text.replace(']','', -1)
    text = text.replace("\r", " ")
    text = text.replace("\n", " ")
    text = text.replace("\t", " ")
    text = text.replace("&nbsp", " ")
    text = text.replace("    ", " ")
    output_post['text'] = re.sub(r'\(http\S+\)', '', text, flags=re.MULTILINE)
    return output_post

def save_post_to_json(path, post):
    name = 'post' + str(post['id']) + '.json'
    with open(join(path, name), 'w', encoding='utf-8') as output:
        json.dump(post, output, indent=4, ensure_ascii=False)
        
def cleanhtml(raw_html):
    cleanr = re.compile('<.*?>|&([a-z0-9]+|#[0-9]{1,6}|#x[0-9a-f]{1,6});')
    cleantext = re.sub(cleanr, '', raw_html)
    return cleantext

bar = progressbar.ProgressBar(maxval=3950).start()
# 3290 - 59400 offset
for offset in range(3200, 3950, 50):
    req_url = 'https://api.vc.ru/v1.8/timeline/index/recent?count=50&offset=' + str(offset)
    request = requests.get(req_url)
    data = json.loads(request.text)
    bar.update(offset)
    for item in data['result']:
        post = parse_post(item)
        date = datetime.fromtimestamp(post['date'])
        if date.year < 2020 and date.year > 2013 and len(post['text']) > 400:
            save_post_to_json(join('D:\\some stuff\\fin_and_inv', str(date.year)), post)
    
bar.finish()

#req_url = 'https://api.vc.ru/v1.8/timeline/index/recent?count=1&offset=' + str(57000)
req_url = 'https://api.vc.ru/v1.8/entry/53101'
request = requests.get(req_url)

data = json.loads(request.text)
item = data['result']
post = parse_post(item)
save_post_to_json(join('D:\\some stuff\\fin_and_inv'), post)      

with open('D:\\some stuff\\fin_and_inv\\2019\\post96660.json', encoding='utf-8') as f:
    r = json.load(f)
    print(r)
