==> 02_data_creation.ipynb <==
import pandas as pd
import json
from os.path import join
from os import listdir
import altair as alt
import pickle

folder_names = ['fintech', 'investment', 'other']

rows_fin = []
rows_inv = []
for folder in folder_names:
    path = join('data', folder)
    posts_names = listdir(path)
    for post in posts_names:
        with open(join(path, post), encoding='utf-8') as f:
            if folder == 'fintech':
                r = json.load(f)
                r['fintech'] = 1
                rows_fin.append(r)
            elif folder == 'investment':
                v = json.load(f)
                v['investment'] = 1
                rows_inv.append(v)
            else:
                q = json.load(f)
                r = dict(q)
                v = dict(q)
                r['fintech'] = 0
                rows_fin.append(r)
                v['investment'] = 0
                rows_inv.append(v)
            
            
df_fin = pd.DataFrame.from_dict(rows_fin, orient='columns')
df_inv = pd.DataFrame.from_dict(rows_inv, orient='columns')

df_fin.head()

df_inv.head()

print('Number of fintech articles: ', df_fin.groupby('fintech').size()[1])
print('Number of investments articles: ', df_inv.groupby('investment').size()[1])
print('Number of other articles: ', df_inv.groupby('investment').size()[0])

with open('pickles\\raw_dataset_inv.pickle', 'wb') as output:
    pickle.dump(df_inv, output)
    
with open('pickles\\raw_dataset_fin.pickle', 'wb') as output:
    pickle.dump(df_fin, output)
