==> 03_feature_engineegring.ipynb <==
import pickle
import os
import requests
from pathlib import Path
import nltk
from nltk import sent_tokenize, word_tokenize, regexp_tokenize
from nltk.corpus import stopwords
import pymorphy2
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer


path_df_fin = "pickles/raw_dataset_fin.pickle"
path_df_inv = "pickles/raw_dataset_inv.pickle"

with open(path_df_fin, 'rb') as data:
    df_fin = pickle.load(data)
    
with open(path_df_inv, 'rb') as data:
    df_inv = pickle.load(data)

df_fin.head()

df_fin.loc[0]['text']

url_stopwords_ru = "https://raw.githubusercontent.com/stopwords-iso/stopwords-ru/master/stopwords-ru.txt"


def get_text(url, encoding='utf-8', to_lower=True):
    url = str(url)
    if url.startswith('http'):
        r = requests.get(url)
        if not r.ok:
            r.raise_for_status()
        return r.text.lower() if to_lower else r.text
    elif os.path.exists(url):
        with open(url, encoding=encoding) as f:
            return f.read().lower() if to_lower else f.read()
    else:
        raise Exception('parameter [url] can be either URL or a filename')


def normalize_tokens(tokens):
    morph = pymorphy2.MorphAnalyzer()
    return [morph.parse(tok)[0].normal_form for tok in tokens]


def remove_stopwords(tokens, stopwords=None, min_length=4):
    if not stopwords:
        return tokens
    stopwords = set(stopwords)
    tokens = [tok
              for tok in tokens
              if tok not in stopwords and len(tok) >= min_length]
    return tokens


def tokenize_n_lemmatize(
    text, stopwords=None, normalize=True, 
    regexp=r'(?u)\b\w{4,}\b'):
    words = [w for sent in sent_tokenize(text)
             for w in regexp_tokenize(sent, regexp)]
    if normalize:
        words = normalize_tokens(words)
    if stopwords:
        words = remove_stopwords(words, stopwords)
    return words

stopwords_ru = get_text(url_stopwords_ru).splitlines()

#text = df_fin.loc[0]['text']
#words = tokenize_n_lemmatize(text, stopwords=stopwords_ru)


# add progress bar
def proc_df(df):
    nrows = len(df)
    text_parsed_list = []
    for row in range(0, nrows):
        text = df.loc[row]['title'] + ' ' + df.loc[row]['text']
        text = text.replace('_', ' ')
        text_parsed_list.append(' '.join(tokenize_n_lemmatize(text)))
    df['text_parsed'] = text_parsed_list

proc_df(df_fin)

proc_df(df_inv)

df_fin.head()

df_inv.head()

X_train_inv, X_test_inv, y_train_inv, y_test_inv = train_test_split(df_inv['text_parsed'], 
                                                    df_inv['investment'], 
                                                    test_size=0.15, 
                                                    random_state=8)

X_train_fin, X_test_fin, y_train_fin, y_test_fin = train_test_split(df_fin['text_parsed'], 
                                                    df_fin['fintech'], 
                                                    test_size=0.15, 
                                                    random_state=8)

with open('Pickles/tfidf.pickle', 'rb') as data:
    tfidf = pickle.load(data)

def to_tfidf(X_train, y_train, X_test, y_test):
    features_train = tfidf.transform(X_train).toarray()
    labels_train = y_train
    print(features_train.shape)

    features_test = tfidf.transform(X_test).toarray()
    labels_test = y_test
    print(features_test.shape)
    
    return features_train, labels_train, features_test, labels_test

features_train_fin, labels_train_fin, features_test_fin, labels_test_fin = to_tfidf(X_train_fin, 
                                                                                   y_train_fin, 
                                                                                   X_test_fin, 
                                                                                   y_test_fin)
print('\n')

features_train_inv, labels_train_inv, features_test_inv, labels_test_inv = to_tfidf(X_train_inv, 
                                                                                   y_train_inv, 
                                                                                   X_test_inv, 
                                                                                   y_test_inv)

from sklearn.feature_selection import chi2
import numpy as np
def check_most_corr_words(features_train, labels_train, category_codes):
    for Product, category_id in sorted(category_codes.items()):
        features_chi2 = chi2(features_train, labels_train == category_id)
        indices = np.argsort(features_chi2[0])
        feature_names = np.array(tfidf.get_feature_names())[indices]
        unigrams = [v for v in feature_names if len(v.split(' ')) == 1]
        print("# '{}' category:".format(Product))
        print("  . Most correlated unigrams:\n. {}".format('\n. '.join(unigrams[-10:])))
        print("")

check_most_corr_words(features_train_fin, labels_train_fin, {'fin': 1, 'not_fin': 0})

check_most_corr_words(features_train_inv, labels_train_inv, {'inv': 1, 'not_inv': 0})

# X_train
with open('Pickles/X_train_inv.pickle', 'wb') as output:
    pickle.dump(X_train_inv, output)
    
# X_test    
with open('Pickles/X_test_inv.pickle', 'wb') as output:
    pickle.dump(X_test_inv, output)
    
# y_train
with open('Pickles/y_train_inv.pickle', 'wb') as output:
    pickle.dump(y_train_inv, output)
    
# y_test
with open('Pickles/y_test_inv.pickle', 'wb') as output:
    pickle.dump(y_test_inv, output)
    
# features_train
with open('Pickles/features_train_inv.pickle', 'wb') as output:
    pickle.dump(features_train_inv, output)

# labels_train
with open('Pickles/labels_train_inv.pickle', 'wb') as output:
    pickle.dump(labels_train_inv, output)

# features_test
with open('Pickles/features_test_inv.pickle', 'wb') as output:
    pickle.dump(features_test_inv, output)

# labels_test
with open('Pickles/labels_test_inv.pickle', 'wb') as output:
    pickle.dump(labels_test_inv, output)

# X_train
with open('Pickles/X_train_fin.pickle', 'wb') as output:
    pickle.dump(X_train_fin, output)
    
# X_test    
with open('Pickles/X_test_fin.pickle', 'wb') as output:
    pickle.dump(X_test_fin, output)
    
# y_train
with open('Pickles/y_train_fin.pickle', 'wb') as output:
    pickle.dump(y_train_fin, output)
    
# y_test
with open('Pickles/y_test_fin.pickle', 'wb') as output:
    pickle.dump(y_test_fin, output)
    
# features_train
with open('Pickles/features_train_fin.pickle', 'wb') as output:
    pickle.dump(features_train_fin, output)

# labels_train
with open('Pickles/labels_train_fin.pickle', 'wb') as output:
    pickle.dump(labels_train_fin, output)

# features_test
with open('Pickles/features_test_fin.pickle', 'wb') as output:
    pickle.dump(features_test_fin, output)

# labels_test
with open('Pickles/labels_test_fin.pickle', 'wb') as output:
    pickle.dump(labels_test_fin, output)
