==> 04_model_trainig_rfc.ipynb <==
import pickle
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.ensemble import RandomForestClassifier
from pprint import pprint
from sklearn.model_selection import RandomizedSearchCV
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score
from sklearn.model_selection import ShuffleSplit
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd

path_df_fin = "pickles/raw_dataset_fin.pickle"
path_df_inv = "pickles/raw_dataset_inv.pickle"

with open(path_df_fin, 'rb') as data:
    df_fin = pickle.load(data)
    
with open(path_df_inv, 'rb') as data:
    df_inv = pickle.load(data)

# X_train
with open('pickles/X_train_inv.pickle', 'rb') as data:
    X_train_inv = pickle.load(data)
    
# X_test    
with open('pickles/X_test_inv.pickle', 'rb') as data:
    X_test_inv = pickle.load(data)
    
# y_train
with open('pickles/y_train_inv.pickle', 'rb') as data:
    y_train_inv = pickle.load(data)
    
# y_test
with open('pickles/y_test_inv.pickle', 'rb') as data:
    y_test_inv = pickle.load(data)
    
# features_train
with open('pickles/features_train_inv.pickle', 'rb') as data:
    features_train_inv = pickle.load(data)

# labels_train
with open('pickles/labels_train_inv.pickle', 'rb') as data:
    labels_train_inv = pickle.load(data)

# features_test
with open('pickles/features_test_inv.pickle', 'rb') as data:
    features_test_inv = pickle.load(data)

# labels_test
with open('pickles/labels_test_inv.pickle', 'rb') as data:
    labels_test_inv = pickle.load(data)

# X_train
with open('pickles/X_train_fin.pickle', 'rb') as data:
    X_train_fin = pickle.load(data)
    
# X_test    
with open('pickles/X_test_fin.pickle', 'rb') as data:
    X_test_fin = pickle.load(data)
    
# y_train
with open('pickles/y_train_fin.pickle', 'rb') as data:
    y_train_fin = pickle.load(data)
    
# y_test
with open('pickles/y_test_fin.pickle', 'rb') as data:
    y_test_fin = pickle.load(data)
    
# features_train
with open('pickles/features_train_fin.pickle', 'rb') as data:
    features_train_fin = pickle.load(data)

# labels_train
with open('pickles/labels_train_fin.pickle', 'rb') as data:
    labels_train_fin = pickle.load(data)

# features_test
with open('pickles/features_test_fin.pickle', 'rb') as data:
    features_test_fin = pickle.load(data)

# labels_test
with open('pickles/labels_test_fin.pickle', 'rb') as data:
    labels_test_fin = pickle.load(data)

# TF-IDF object
with open('pickles/tfidf.pickle', 'rb') as data:
    tfidf = pickle.load(data)

print(features_train_fin.shape)
print(features_test_fin.shape)
print('\n')
print(features_train_inv.shape)
print(features_test_inv.shape)

# FINTECH

# Create the parameter grid based on the results of random search 
bootstrap = [False]
max_depth = [3,4 ,5,6, 7, 8, 10, 15]
max_features = ['sqrt']
min_samples_leaf = [1, 2]
min_samples_split = [2, 3, 5]
n_estimators = [800]

param_grid = {
    'bootstrap': bootstrap,
    'max_depth': max_depth,
    'max_features': max_features,
    'min_samples_leaf': min_samples_leaf,
    'min_samples_split': min_samples_split,
    'n_estimators': n_estimators
}

# Create a base model
rfc = RandomForestClassifier(random_state=8)

# Manually create the splits in CV in order to be able to fix a random_state (GridSearchCV doesn't have that argument)
cv_sets = ShuffleSplit(n_splits = 3, test_size = .33, random_state = 8)

# Instantiate the grid search model
grid_search_fin = GridSearchCV(estimator=rfc, 
                           param_grid=param_grid,
                           scoring='accuracy',
                           cv=cv_sets,
                           verbose=1)

# Fit the grid search to the data
grid_search_fin.fit(features_train_fin, labels_train_fin)

print("The best hyperparameters from Grid Search are:")
print(grid_search_fin.best_params_)
print("")
print("The mean accuracy of a model with these hyperparameters is:")
print(grid_search_fin.best_score_)

best_rfc_fin = grid_search_fin.best_estimator_

best_rfc_fin

best_rfc_fin.fit(features_train_fin, labels_train_fin)

rfc_pred_fin = best_rfc_fin.predict(features_test_fin)

# Training accuracy
print("The training accuracy is: ")
print(accuracy_score(labels_train_fin, best_rfc_fin.predict(features_train_fin)))

# Test accuracy
print("The test accuracy is: ")
print(accuracy_score(labels_test_fin, rfc_pred_fin))

# INVESTMENT

# Create the parameter grid based on the results of random search 
bootstrap = [False]
max_depth = [3, 4, 5, 6, 7, 8, 9, 10, 15]
max_features = ['sqrt']
min_samples_leaf = [1, 2, 4]
min_samples_split = [2, 3, 5]
n_estimators = [800]

param_grid = {
    'bootstrap': bootstrap,
    'max_depth': max_depth,
    'max_features': max_features,
    'min_samples_leaf': min_samples_leaf,
    'min_samples_split': min_samples_split,
    'n_estimators': n_estimators
}

# Create a base model
rfc = RandomForestClassifier(random_state=8)

# Manually create the splits in CV in order to be able to fix a random_state (GridSearchCV doesn't have that argument)
cv_sets = ShuffleSplit(n_splits = 3, test_size = .33, random_state = 8)

# Instantiate the grid search model
grid_search_inv = GridSearchCV(estimator=rfc, 
                           param_grid=param_grid,
                           scoring='accuracy',
                           cv=cv_sets,
                           verbose=1)

# Fit the grid search to the data
grid_search_inv.fit(features_train_inv, labels_train_inv)

print("The best hyperparameters from Grid Search are:")
print(grid_search_inv.best_params_)
print("")
print("The mean accuracy of a model with these hyperparameters is:")
print(grid_search_inv.best_score_)

best_rfc_inv = grid_search_inv.best_estimator_

best_rfc_inv

best_rfc_inv.fit(features_train_inv, labels_train_inv)

rfc_pred_inv = best_rfc_inv.predict(features_test_inv)

# Training accuracy
print("The training accuracy is: ")
print(accuracy_score(labels_train_inv, best_rfc_inv.predict(features_train_inv)))

# Test accuracy
print("The test accuracy is: ")
print(accuracy_score(labels_test_inv, rfc_pred_inv))

with open('models/best_rfc_fin.pickle', 'wb') as output:
    pickle.dump(best_rfc_fin, output)

with open('models/best_rfc_inv.pickle', 'wb') as output:
    pickle.dump(best_rfc_inv, output)
