==> 05_report_3.ipynb <==
import pickle
import pandas as pd
import numpy as np
from datetime import datetime

with open('pickles/parsed_big_df.pickle', 'rb') as data:
    df = pickle.load(data)

with open('models/best_rfc_fin.pickle', 'rb') as data:
    rfc_fin = pickle.load(data)
    
with open('models/best_rfc_inv.pickle', 'rb') as data:
    rfc_inv = pickle.load(data)

df.head()

# Parameter election
ngram_range = (1, 1)
min_df = 10
max_df = 1.
max_features = 8000

tfidf = TfidfVectorizer(encoding='utf-8',
                        ngram_range=ngram_range,
                        stop_words=None,
                        lowercase=False,
                        max_df=max_df,
                        min_df=min_df,
                        max_features=max_features,
                        norm='l2',
                        sublinear_tf=True)

features = tfidf.fit_transform(df['text_parsed']).toarray()

# TF-IDF object
with open('pickles/tfidf.pickle', 'wb') as output:
    pickle.dump(tfidf, output)

fin_pred = rfc_fin.predict(features)

inv_pred = rfc_inv.predict(features)

np.unique(fin_pred, return_counts=True)

np.unique(inv_pred, return_counts=True)

tfidf

fin = [0] * 72
inv = [0] * 72
total = [0] * 72
for i in range(len(df)):
    date = datetime.fromtimestamp(df.loc[i]['date'])
    j = (date.year - 2014) * 12 + date.month - 1 
    total[j] += 1
    if fin_pred[i] == 1:
        fin[j] += 1
    if inv_pred[i] == 1:
        inv[j] += 1        

data = {'Total' : total, 'Fintech' : fin, 'Investment' : inv} 

out = pd.DataFrame(data) 

out.head()

out.to_csv('final.csv', sep=';')

tot_res = [0] * 72
fin_res = [0] * 72
inv_res = [0] * 72
words_tf_fin = ['финтех', 'криптовалюта', 'блокчейн', 'краудфандинг']
words_tf_fin_p = [('инвестиция','стартап'),('инвестиция', 'финтех'),('онлайн', 'банк'), ('финансовый', 'рынок')]
words_tf_inv = [('венчурный', 'капитал'), ('венчурный', 'финансирование')]
for i in range(len(df)):
    date = datetime.fromtimestamp(df.loc[i]['date'])
    j = (date.year - 2014) * 12 + date.month - 1 
    tot_res[j] += 1
    f = False
    for w in words_tf_fin:
        if df.loc[i]['text_parsed'].find(w) != -1:
            f = True
            fin_res[j] += 1
            break
    if f == False:    
        for w1, w2 in words_tf_fin_p:
            if df.loc[i]['text_parsed'].find(w1) != -1 and df.loc[i]['text_parsed'].find(w2) != -1:
                fin_res[j] += 1
                break
    for w1, w2 in words_tf_inv:
            if df.loc[i]['text_parsed'].find(w1) != -1 and df.loc[i]['text_parsed'].find(w2) != -1:
                inv_res[j] += 1
                break

data = {'Total' : tot_res, 'Fintech' : fin_res, 'Investment': inv_res} 

out = pd.DataFrame(data) 

out.to_csv('vc_key_words.csv', sep=';')
